+++ CHANGELOG.md
@@ -12,6 +12,12 @@ All changes to the MarkLogic (backend) portion of LUX capable of impacting the r
 ### Changed
 - Changed the relative paths to absolute paths for instances of environmentLib.mjs
 
+### Added
+- Added Collection and Responsible Unit facets for Collection results ([#461](https://github.com/project-lux/lux-marklogic/issues/461))
+
+### Removed
+- Removed the Part Of facet for Collection results ([#461](https://github.com/project-lux/lux-marklogic/issues/461))
+
 ## v4.1.0 - 2026-08-17
 ### Removed
 - Removed unused indexes, this frees up some memory and disk space ([#652](https://github.com/project-lux/lux-marklogic/issues/652))